requirements.txt
opencv-python==4.10.0.84
numpy==1.26.4
ultralytics==8.3.0
torch==2.4.1
torchvision==0.19.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pyyaml==6.0.1
matplotlib==3.9.2
requests==2.32.3
